Allow elliptic curve keys in from_cryptography_key().
Elliptic curve support in PKey objects still needs to be implemented.

With this change an (Certificate Transparency) SCT can be verified
against the public key of a CT-Log. CT-Logs usually are signed with an
elliptic curve digest.  The argument `cert` of the function
`verify(cert, signature, data, digest)` is just a wrapper for its `pkey`
attribute.  This attribute now can contain an ec-pubkey.  Here is an example
of this usage:
https://github.com/theno/ctutlz/blob/master/ctutlz/sct/verification.py#L29