Fix various issues around X509_STORE_CTX reuse (#1272)
* Don't try to access X509_STORE_CTX after _cleanup

Although OpenSSL happens to leave the errors in there on
X509_STORE_CTX_cleanup, in no other OpenSSL API is accessing a cleaned
up object meaningful. Do it in the other order.

* Internal variables are internal

The underscore-prefixed variables were not intended to be exposed as
public API, so don't bother exposing it in the first place.

* Don't reuse X509_STORE_CTXs

There's a lot of history with X509_STORE_CTX's somewhat messy transition
from a stack-allocated type to a heap-allocated type. (This is why a
double X509_STORE_CTX_init used to leak memory.) We can avoid all this
mess by just making a new X509_STORE_CTX each time.