Set SSL_MODE_ACCEPT_MOVING_WRITE_BUFFER in calling OpenSSL (#1287)
* Squashed commit: Fix for problem caused by SSL_WANT_READ or SSL_WANT_WRITE errors.

When SSL_WANT_READ or SSL_WANT_WRITE are encountered, it's typical to retry the call but this must be repeated with the exact same arguments. Without this change, openSSL requires that the address of the buffer passed is the same. However, buffers in python can change location in some circumstances which cause the retry to fail.  By add the setting SSL_MODE_ACCEPT_MOVING_WRITE_BUFFER, the requirement for the same buffer address is forgiven and the retry has a better chance of success.  See cherrypy/cheroot#245 for discussion.

* removed remaining logger statements

* removed erroneous typo in ci.yml

* small fixes

* fixed types

* missing changes

* Revert "fixed types"

This reverts commit bb1d143da2ba4d9a21f25595b34ac38dcba9213c.

* format fix