Add methods related to client hello callback
These methods allow a server to switch the context when looking at ALPN
and servername together. This is for example require when implementing the
ACME tls-alpn/1 protocol.

Unfortunately, OpenSSL does not provide any utility function to actually
parse ClientHello extensions when using these APIs and the user has to write
their own methods.

Closes: pyca/pyopenssl#1430
Signed-off-by: Arne Schwabe <arne@rfc2549.org>